requirements.txt
aiogram==2.25.1
transformers==4.38.2
torch>=2.0.0
accelerate>=0.27.0
python-dotenv>=1.0.0
sentencepiece>=0.1.99
